Refactor NavBar component and userSlice***
This commit refactors the NavBar component by adding the useDispatch hook and setUser action from the userSlice. It also updates the useEffect hook to set the loggedInUser state correctly. Additionally, the commit removes unnecessary console.log statements.

In the userSlice, the commit removes the isLoggedIn property from the initialState and the state object. This property is no longer needed.

No changes were made to the home and login pages.

In the user.controllers.js file, the commit removes unnecessary fields from the select query in the loginUser function.